requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.25.1
selenium==4.17.2
webdriver-manager==4.0.1
firebase-admin==6.4.0
